add auth routes and auth controller and utils with generate token
