--- routes/commonRoutes.js
@@ -1,9 +1,0 @@
-const express = require('express');
-
-const router = express.Router();
-const commonController = require('../controllers/commonController');
-
-router.route('/validate-token')
-  .post(commonController.validateToken)
-
-module.exports = router;
