+++ routes/index.js
@@ -1,17 +1,17 @@
 const express = require('express');
 const path = require('path');
 const usersRouter = require('./userRoutes');
-const commonRouter = require('./commonRoutes');
 const rootRouter = require('./root');
 const customerRouter = require('./customerRoutes');
+const authRouter = require('./authRoutes');
 
 const routerApi = (app) => {
   const router = express.Router();
   app.use('/api/v1', router);
   router.use('/users', usersRouter);
-  router.use('/common', commonRouter);
+  router.use('/customer', customerRouter);
+  router.use('/auth', authRouter);
   router.use('/', rootRouter);
-  router.use('/', customerRouter);
   router.use('/', express.static(path.join(__dirname, 'public')));
 
   router.all('*', (req, res) => {
