+++ utils/utils.js
@@ -1,0 +1,9 @@
+const jwt = require('jsonwebtoken');
+
+const generateToken = (id, role) => {
+  return jwt.sign({ id, role: role}, process.env.JWT_SECRET, { expiresIn: '30d' });
+};
+
+module.exports = {
+  generateToken,
+}
